requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-Migrate
psycopg2-binary
pandas
gunicorn
